requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python
keras
torch
torchvision
timm
transformers
tensorflow
tqdm
scikit-learn
ultralytics
pandas
matplotlib
seaborn
